+++ R/geom-measures.R
@@ -128,10 +128,7 @@ st_perimeter = function(x, ...) {
 		if (!requireNamespace("lwgeom", quietly = TRUE))
 			stop("package lwgeom required, please install it first")
 		# note that units are handled appropriately by lwgeom
-		if (packageVersion("lwgeom") <= "0.2-13")
-			lwgeom::st_perimeter(x)
-		else
-			lwgeom::st_perimeter_lwgeom(x)
+		lwgeom::st_perimeter_lwgeom(x)
 	}
 }