+++ R/init.R
@@ -5,7 +5,6 @@
 #' @importFrom grid convertHeight convertUnit convertWidth current.viewport linesGrob nullGrob pathGrob pointsGrob polylineGrob unit viewport
 #' @import graphics
 #' @importFrom grDevices dev.size rgb cm
-#' @importFrom Rcpp evalCpp
 #' @importFrom DBI dbConnect dbDisconnect dbExecute dbGetQuery dbReadTable dbSendQuery dbWriteTable
 #' @importFrom units as_units drop_units make_unit_label set_units
 #' @importFrom classInt classIntervals
@@ -27,6 +26,7 @@ setOldClass(c("sfc_GEOMETRY", "sfc"))
 setOldClass(c("sfc_GEOMETRYCOLLECTION", "sfc"))
 setOldClass("sfg")
 setOldClass("crs")
+setOldClass("bbox")
 
 .sf_cache <- new.env(FALSE, parent=globalenv())